+++ src/component/page/veileder/MineDeltakereTabell.tsx
@@ -1,7 +1,7 @@
 import { Table } from '@navikt/ds-react'
 import React, { useEffect, useState } from 'react'
 
-import { useTiltaksoversiktSokContext } from '../../../store/TiltaksoversiktSokProvider'
+import { useTiltaksoversiktSokContext } from '../../../store/TiltaksoversiktSokContextProvider'
 import { finnNesteSortering } from '../../../utils/sortering-utils'
 import globalStyles from '../../../globals.module.scss'
 import { sorterVeiledersDeltakere } from './sortering'
@@ -10,7 +10,7 @@ import { IngenDeltakere } from './ingen-deltakere/IngenDeltakere'
 import { TabellHeaderVeileder } from './TabellHeaderVeileder'
 import { TabellBodyVeileder } from './TabellBodyVeileder'
 import { filtrerDeltakerliste } from '../../../utils/filtrering-utils'
-import { useVeilederFilterMenyStore } from './store/veileder-filter-meny-store-provider'
+import { useVeilederFilterContext } from './store/VeilederFilterContextProvider'
 import { AlertInfoMessage } from '../../felles/alert-info-message/AlertInfoMessage'
 
 interface MineDeltakereTabellProps {
@@ -29,7 +29,7 @@ export const MineDeltakereTabell = (
     hendelseFilter,
     deltakerlisteFilter,
     veiledertypeFilter
-  } = useVeilederFilterMenyStore()
+  } = useVeilederFilterContext()
   const [deltakereBearbeidet, setDeltakereBearbeidet] = useState<
     VeiledersDeltaker[]
   >(sorterVeiledersDeltakere(mineDeltakere, deltakerSortering))
