+++ src/store/store-provider.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
 
 import { TilbakelenkeContextProvider } from './TilbakelenkeContextProvider'
-import { TiltaksoversiktSokProvider } from './TiltaksoversiktSokProvider'
+import { TiltaksoversiktSokContextProvider } from './TiltaksoversiktSokContextProvider'
 import { InnloggetBrukerContextProvider } from './InnloggetBrukerContextProvider'
 import { KoordinatorsDeltakerlisterContextProvider } from './KoordinatorsDeltakerlisterContextProvider'
 import { KoordinatorFilterContextProvider } from '../component/page/deltakerliste-detaljer/store/KoordinatorFilterContextProvider'
-import { VeilederFilterMenyStoreProvider } from '../component/page/veileder/store/veileder-filter-meny-store-provider'
+import { VeilederFilterContextProvider } from '../component/page/veileder/store/VeilederFilterContextProvider'
 
 interface StoreProviderProps {
   children: React.ReactNode
@@ -15,19 +15,19 @@ const StoreProvider = (
   props: StoreProviderProps
 ): React.ReactElement<StoreProviderProps> => {
   return (
-    <TiltaksoversiktSokProvider>
+    <TiltaksoversiktSokContextProvider>
       <KoordinatorsDeltakerlisterContextProvider>
         <KoordinatorFilterContextProvider>
-          <VeilederFilterMenyStoreProvider>
+          <VeilederFilterContextProvider>
             <InnloggetBrukerContextProvider>
               <TilbakelenkeContextProvider>
                 {props.children}
               </TilbakelenkeContextProvider>
             </InnloggetBrukerContextProvider>
-          </VeilederFilterMenyStoreProvider>
+          </VeilederFilterContextProvider>
         </KoordinatorFilterContextProvider>
       </KoordinatorsDeltakerlisterContextProvider>
-    </TiltaksoversiktSokProvider>
+    </TiltaksoversiktSokContextProvider>
   )
 }
